Changed NewOptionalComponent and NewOptionalBox into OptionalComponent.svelte and OptionalBox.ts
