+++ packages/core/src/editor/boxes/BoxFactory.ts
@@ -17,17 +17,15 @@ import {
     HorizontalLayoutBox,
     VerticalLayoutBox,
     TableCellBox,
-    OptionalBox,
     LimitedControlBox,
     ButtonBox,
     NumberDisplay,
     type AbstractExternalBox,
     PartListReplacerBox,
     isPartListReplacerBox,
-    ReferenceBox, MultiLineTextBox, NewOptionalBox
+    ReferenceBox, MultiLineTextBox, OptionalBox
 } from "./internal.js"
 import type { SelectOption } from "./internal.js";
-import type { BoolFunctie } from "./internal.js";
 
 type RoleCache<T extends Box> = {
     [role: string]: T;
@@ -50,7 +48,6 @@ let selectCache: BoxCache<SelectBox> = {};
 let referenceCache: BoxCache<ReferenceBox> = {};
 // let indentCache: BoxCache<IndentBox> = {};
 let optionalCache: BoxCache<OptionalBox> = {};
-let newOptionalCache: BoxCache<NewOptionalBox> = {};
 // let svgCache: BoxCache<SvgBox> = {};
 let horizontalLayoutCache: BoxCache<HorizontalLayoutBox> = {};
 let verticalLayoutCache: BoxCache<VerticalLayoutBox> = {};
@@ -92,7 +89,6 @@ export class BoxFactory {
         selectCache = {};
         referenceCache = {};
         optionalCache = {};
-        newOptionalCache = {};
         horizontalLayoutCache = {};
         verticalLayoutCache = {};
         horizontalListCache = {};
@@ -494,13 +490,34 @@ export class BoxFactory {
         return result;
     }
 
+    // static optional(
+    //     node: FreNode,
+    //     role: string,
+    //     condition: BoolFunctie,
+    //     box: Box,
+    //     mustShow: boolean,
+    //     optional: Box,
+    //     initializer?: Partial<OldOptionalBox>,
+    // ): OldOptionalBox {
+    //     // TODO This only works with cache on, should also work with cache off.
+    //     // if (cacheOptionalOff) {
+    //     //     return new OptionalBox(element, role, condition, box, mustShow, actionText);
+    //     // }
+    //     // 1. Create the optional box, or find the one that already exists for this element and role
+    //     const creator = () => new OldOptionalBox(node, role, condition, box, mustShow, optional);
+    //     const result: OldOptionalBox = this.find<OldOptionalBox>(node, role, creator, optionalCache);
+    //
+    //     // 2. Apply the other arguments in case they have changed
+    //     FreUtils.initializeObject(result, initializer);
+    //
+    //     return result;
+    // }
+
     static optional(
         node: FreNode,
         role: string,
-        condition: BoolFunctie,
-        box: Box,
-        mustShow: boolean,
-        optional: Box,
+        placeHolder: string,
+        contentBox: Box,
         initializer?: Partial<OptionalBox>,
     ): OptionalBox {
         // TODO This only works with cache on, should also work with cache off.
@@ -508,29 +525,8 @@ export class BoxFactory {
         //     return new OptionalBox(element, role, condition, box, mustShow, actionText);
         // }
         // 1. Create the optional box, or find the one that already exists for this element and role
-        const creator = () => new OptionalBox(node, role, condition, box, mustShow, optional);
+        const creator = () => new OptionalBox(node, role, placeHolder, contentBox);
         const result: OptionalBox = this.find<OptionalBox>(node, role, creator, optionalCache);
-
-        // 2. Apply the other arguments in case they have changed
-        FreUtils.initializeObject(result, initializer);
-
-        return result;
-    }
-
-    static newOptional(
-        node: FreNode,
-        role: string,
-        placeHolder: string,
-        contentBox: Box,
-        initializer?: Partial<NewOptionalBox>,
-    ): NewOptionalBox {
-        // TODO This only works with cache on, should also work with cache off.
-        // if (cacheOptionalOff) {
-        //     return new OptionalBox(element, role, condition, box, mustShow, actionText);
-        // }
-        // 1. Create the optional box, or find the one that already exists for this element and role
-        const creator = () => new NewOptionalBox(node, role, placeHolder, contentBox);
-        const result: NewOptionalBox = this.find<NewOptionalBox>(node, role, creator, newOptionalCache);
 
         // 2. Apply the other arguments in case they have changed
         FreUtils.initializeObject(result, initializer);
