+++ packages/core/src/editor/boxes/OptionalBox.ts
@@ -1,102 +1,239 @@
-import { autorun } from "mobx";
-import type { FreNode } from "../../ast/index.js";
-import { FreUtils } from "../../util/index.js";
-import { Box } from "./internal.js";
+import { Box } from "./Box.js"
+import type { SelectOption } from "./SelectOption.js"
+import type { FreEditor } from "../FreEditor.js"
+import { autorun } from "mobx"
+import type { FreNamedNode, FreNode, FreNodeReference } from "../../ast/index.js"
+import { FreUtils, isNullOrUndefined } from "../../util/index.js"
+import { BehaviorExecutionResult, createOptions, executeSingleBehavior } from "../util/index.js"
+import { FreLogger } from "../../logging/index.js"
+import { FreLanguage, type FreLanguageProperty } from "../../language/index.js"
+import { AST } from "../../change-manager/index.js"
+import { isBooleanControlBox } from "./BooleanControlBox.js"
+import { isSelectBox } from "./SelectBox.js"
 
-export type BoolFunctie = () => boolean;
+const LOGGER: FreLogger = new FreLogger("OptionalBox");
 
-/**
- * OptionalBox holds the content from a projection that is optional. This content is always present in the
- * attribute 'content'. Next to the content there is a 'placeholder' box, which is shown when the content is not
- * present in the FreNode model.
- * The attributes 'mustShow' and 'condition' determine which of the pair [content, placeholder] is shown. If the 'condition'
- * results in true, then the content box is shown. If 'mustShow' is true, then the content box is also shown, even though
- * there may not be actual content within the FreNode model. The latter is set by the custom action, that is coupled
- * to this OptionalBox, which is triggered by the user.
- */
 export class OptionalBox extends Box {
-    readonly kind = "OptionalBox";
+    readonly kind: string = "OptionalBox";
 
     content: Box = null;
-    placeholder: Box = null;
-    _mustShow: boolean = false; // is set to true by action that does not (yet) change the model, but causes part of the optional to be shown
-    condition: () => boolean; // a condition based on the model that determines whether the optional is shown
+    placeholder: string = 'placeholder';
+    /**
+     * The name of the concept of the node attached to this box.
+     */
+    nodeConcept: string;
+    /**
+     * Filled with the name of the concept, in case the optional needs to create new concept instance or reference.
+     * May remain undefined when the property is primitive.
+     */
+    conceptOfProperty?: string;
+    /**
+     * Holds the definition of the optional property that is contained in this box.
+     */
+    private _propDef: FreLanguageProperty;
 
-    get mustShow() {
-        return this._mustShow;
-    }
-    set mustShow(v: boolean) {
-        this._mustShow = v;
-        this.isDirty();
-    }
-
-    constructor(
-        node: FreNode,
-        role: string,
-        condition: BoolFunctie,
-        box: Box,
-        mustShow: boolean,
-        placeholder: Box,
-        initializer?: Partial<OptionalBox>,
+    constructor(node: FreNode,
+                role: string,
+                placeHolder: string,
+                contentBox: Box,
+                initializer?: Partial<OptionalBox>,
     ) {
         super(node, role);
         FreUtils.initializeObject(this, initializer);
-        this.content = box;
-        box.parent = this;
-        // TODO question: should not the role be diff from role of this box? Where is the "action" prefix added?
-        this.placeholder = placeholder;
-        this.placeholder.parent = this;
-        this.mustShow = mustShow;
-        this.condition = condition;
-        autorun(this.conditionChanged);
+        this.placeholder = placeHolder
+        this.content = contentBox;
+        contentBox.parent = this;
+        autorun(this.contentChanged);
+        this.nodeConcept = this.node.freLanguageConcept();
     }
 
-    get showByCondition(): boolean {
-        return this.condition();
+    getOptions(editor: FreEditor): SelectOption[] {
+        LOGGER.log(`getOptions: ${this.nodeConcept}, this.propDef: ${this.propDef?.propertyKind}`);
+        if (this.propDef.propertyKind === "primitive") {
+            if (this.propDef.type === "boolean") {
+                return this.getBooleanOptions(editor);
+            }
+            return [];
+        } else {
+            LOGGER.log("getOptions for " + this.$id + "- " + this.conceptOfProperty + "." + this.propertyName)
+            return createOptions(editor, this.node, this, this.conceptOfProperty)
+        }
+    }
+
+    executeOption(editor: FreEditor, option: SelectOption): BehaviorExecutionResult {
+        console.log("OptionalBox executeOption " + JSON.stringify(option) + ` ${this.nodeConcept}, this.propDef: ${this.propDef?.propertyKind}`);
+        if (this.propDef.propertyKind === "primitive") {
+            const self: OptionalBox = this;
+            console.log(`isPrimitive for property ${this.propertyName}, actualValue: "${self.node[self.propertyName]}"`);
+            if (this.propDef.type === "boolean") {
+                const boolBox: Box = this.content.firstEditableChild;
+                console.log(`getOptions: boolBox: ${boolBox?.kind}`);
+                if (isBooleanControlBox(boolBox)) {
+                    AST.change(() => {
+                        if (option.id === boolBox.labels.yes) {
+                            self.node[self.propertyName] = true;
+                        } else if (option.id === boolBox.labels.no) {
+                            self.node[self.propertyName] = false;
+                        } else if (option.id === boolBox.labels.unknown) {
+                            self.node[self.propertyName] = undefined;
+                        }
+                    });
+                } else if (isSelectBox(boolBox)) {
+                    boolBox.executeOption(editor, option);
+                }
+            } else if (this.propDef.type === "string") {
+                console.log("found string");
+                self.node[self.propertyName] = " ";
+            } else if (this.propDef.type === "number") {
+                self.node[self.propertyName] = 1; // todo improve when undefined value for numbers has been decided
+            }
+        } else {
+            FreUtils.CHECK(!!option?.action, `OptionalBox.executeOption: action missing for ${option?.label}`)
+            if (!!option.action) {
+                return executeSingleBehavior(option.action, this, option.label, editor);
+            }
+            LOGGER.log("<== DONE OptionalBox executeOption ");
+        }
+        return BehaviorExecutionResult.NULL;
     }
 
     /**
-     * Ensure a refresh is triggered if the condition for showing this optional box has changed.
+     * Sets the value of the property that is contained in this Optional Box to not present / not set.
      */
-    conditionChanged = () => {
-        console.log("AUTORUN showByCondition, this.mustShow: " + this.mustShow);
-        this.condition();
+    removeContent() {
+        LOGGER.log(`removeContent ${this.id}`);
+        const self: OptionalBox = this;
+        if (this.propDef.propertyKind === "primitive") {
+            if (this.propDef.type === "boolean") {
+                AST.change(() => {
+                    self.node[self.propertyName] = undefined;
+                });
+            } else if (this.propDef.type === "string") {
+                AST.change(() => {
+                    self.node[self.propertyName] = "";
+                });
+            } else if (this.propDef.type === "number") {
+                AST.change(() => {
+                    self.node[self.propertyName] = 0; // todo improve when undefined value for numbers has been decided
+                });
+            }
+        } else if (this.propDef.propertyKind === "part") {
+            AST.change(() => {
+                self.node[self.propertyName] = undefined;
+            });
+        }
+        console.log(`removeContent ${this.id}: "${this.node[this.propertyName]}"`);
         this.isDirty();
-    };
+    }
+
+    /**
+     * Returns true when the property that is contained in this Optional Box is not present / not set.
+     */
+    isEmpty(): boolean {
+        console.log(`isEmpty for property ${this.propertyName} of concept ${this.nodeConcept}`);
+        if (isNullOrUndefined(this.propDef)) {
+            console.log(`Cannot find property definition for property ${this.propertyName} of concept ${this.nodeConcept}.`)
+            return false;
+        }
+        if (this.propDef.propertyKind === "primitive") {
+            const actualValue = this.node[this.propertyName];
+            console.log(`isPrimitive for property ${this.propertyName}, actualValue: "${actualValue}"`);
+            if (this.propDef.type === "string") {
+                if (isNullOrUndefined(actualValue)) {
+                    console.log(`Error in model, property ${this.propertyName} of concept ${this.nodeConcept} is undefined.`)
+                    return false;
+                }
+                return actualValue.length === 0;
+            } else if (this.propDef.type === "boolean") {
+                return isNullOrUndefined(actualValue);
+            } else if (this.propDef.type === "number") {
+                return isNullOrUndefined(actualValue); // todo improve when undefined value for numbers has been decided
+            }
+        } else if (this.propDef.propertyKind === "part") {
+            const actualValue: FreNode[] = FreLanguage.getInstance().getPropertyValue(this.node, this.propDef);
+            return isNullOrUndefined(actualValue) || actualValue.length === 0;
+        } else if (this.propDef.propertyKind === "reference") {
+            const actualValue: FreNodeReference<FreNamedNode>[] = FreLanguage.getInstance().getReferencePropertyValue(this.node, this.propDef);
+            return isNullOrUndefined(actualValue) || actualValue.length === 0;
+        }
+        return true;
+    }
+
+    contentChanged = () => {
+        LOGGER.log(`contentChanged ${this.id}`);
+        this.isEmpty();
+        this.isDirty();
+    }
 
     /**
      * Get the first selectable leaf box in the tree with `this` as root.
      */
     get firstLeaf(): Box {
-        if (this.condition() || this.mustShow) {
+        if (!this.isEmpty()) {
             return this.content.firstLeaf;
-        } else {
-            return this.placeholder;
         }
+        return null;
     }
 
     get lastLeaf(): Box {
-        if (this.condition() || this.mustShow) {
+        if (!this.isEmpty()) {
             return this.content.lastLeaf;
-        } else {
-            return this.placeholder;
         }
+        return null;
     }
 
     get firstEditableChild(): Box {
-        if (this.condition() || this.mustShow) {
+        if (!this.isEmpty()) {
             return this.content.firstEditableChild;
-        } else {
-            return this.placeholder;
         }
+        return null;
     }
 
     get children(): ReadonlyArray<Box> {
-        if (this.condition() || this.mustShow) {
+        if (!this.isEmpty()) {
             return [this.content];
-        } else {
-            return [this.placeholder];
         }
+        return [];
+    }
+
+    /**
+     * Returns the definition of the optional property that is contained in this box.
+     */
+    private get propDef(): FreLanguageProperty {
+        if (isNullOrUndefined(this._propDef)) {
+            this._propDef = FreLanguage.getInstance().classifierProperty(this.nodeConcept, this.propertyName)
+        }
+        return this._propDef;
+    }
+
+    /**
+     * Returns the SelectOptions for the property, when it is of boolean type.
+     * @param editor
+     * @private
+     */
+    private getBooleanOptions(editor: FreEditor): SelectOption[] {
+        const boolBox: Box = this.content.firstEditableChild;
+        console.log(`getOptions: boolBox: ${boolBox?.kind}`);
+        if (isBooleanControlBox(boolBox)) {
+            // Create two options, one for true and one for false
+            return [
+                {
+                    id: boolBox.labels.yes,
+                    label: boolBox.labels.yes,
+                    description: `true value for ${ this.propertyName }`,
+                },
+                {
+                    id: boolBox.labels.no,
+                    label: boolBox.labels.no,
+                    description: `false value for ${ this.propertyName }`,
+                }
+            ];
+        } else if (isSelectBox(boolBox)) {
+            // Use the options from the SelectBox, but filter out the 'unknown' option
+            const rawResult = boolBox.getOptions(editor);
+            return [rawResult[0], rawResult[1]];
+        }
+        return [];
     }
 }
